Move Utils / Exceptions module to main package and update imports
