+++ src/ragas/exceptions.py
@@ -49,3 +49,76 @@ class UploadException(RagasException):
     def __init__(self, status_code: int, message: str):
         self.status_code = status_code
         super().__init__(message)
+
+
+# Exceptions migrated from experimental module
+class RagasError(Exception):
+    """Base class for all Ragas-related exceptions."""
+
+    pass
+
+
+class ValidationError(RagasError):
+    """Raised when field validation fails."""
+
+    pass
+
+
+class DuplicateError(RagasError):
+    """Exception raised when a duplicate resource is created."""
+
+    pass
+
+
+class NotFoundError(RagasError):
+    """Exception raised when a resource is not found."""
+
+    pass
+
+
+class ResourceNotFoundError(NotFoundError):
+    """Exception raised when a resource doesn't exist."""
+
+    pass
+
+
+class ProjectNotFoundError(ResourceNotFoundError):
+    """Exception raised when a project doesn't exist."""
+
+    pass
+
+
+class DatasetNotFoundError(ResourceNotFoundError):
+    """Exception raised when a dataset doesn't exist."""
+
+    pass
+
+
+class ExperimentNotFoundError(ResourceNotFoundError):
+    """Exception raised when an experiment doesn't exist."""
+
+    pass
+
+
+class DuplicateResourceError(RagasError):
+    """Exception raised when multiple resources exist with the same identifier."""
+
+    pass
+
+
+class DuplicateProjectError(DuplicateResourceError):
+    """Exception raised when multiple projects exist with the same name."""
+
+    pass
+
+
+class DuplicateDatasetError(DuplicateResourceError):
+    """Exception raised when multiple datasets exist with the same name."""
+
+    pass
+
+
+class DuplicateExperimentError(DuplicateResourceError):
+    """Exception raised when multiple experiments exist with the same name."""
+
+    pass
